+++ src/BlockHorizons/BlockSniper/brush/types/LayerType.php
@@ -17,9 +17,12 @@ class LayerType extends BaseType{
 
 	public const ID = self::TYPE_LAYER;
 
+	/** @var Vector3 */
+	private $center;
+
 	public function __construct(Brush $brush, ChunkManager $level, \Generator $blocks = null){
 		parent::__construct($brush, $level, $blocks);
-		$this->center = $brush->getPlayer()->getTargetBlock($brush->getPlayer()->getViewDistance() * 16)->asVector3();
+		$this->center = $this->target($brush);
 	}
 
 	/**